pyproject.toml
[tool.poetry]
name = "pinecone"
version = "0.1.0"
description = "Pinecone tutorial"
authors = ["Preethi Prakash <Preethi93@icloud.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pandas = "^2.2.2"
pinecone-client = {extras = ["grpc"], version = "^4.1.0"}
numpy = "^1.26.4"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
